Handle type exception from library code
This gets us a proper exit code even when library code fails and also decorates the error with spinner.fail.

I looked wide and far and found heaps of zabbix implementations that all exhibit this bug. It's only a question of time until i write my own (modern) module (or switch all of this over to python).